Improvements see desc
I created a new file specifically to hold the alarms list, this can then be imported into any other file where we can retrieve the list of alarms (in List<AlarmSettings> data type).
I also changed the code to look at the size of the list of alarms and to take the data from that list in order to dynamically create the container buttons that we are currently utilizing.